+++ source/Room.cpp
@@ -12,8 +12,8 @@ Room::Room()
 void Room::load()
 {
     backgroundSprite.setScale(sf::Vector2f(
-        1.f * LOGICAL_WIDTH / backgroundTexture.getSize().x,
-        1.f * LOGICAL_HEIGHT / backgroundTexture.getSize().y));
+        1.f * LOGICAL_WIDTH / static_cast<float>(backgroundTexture.getSize().x),
+        1.f * LOGICAL_HEIGHT / static_cast<float>(backgroundTexture.getSize().y)));
 }
 
 void Room::draw(sf::RenderWindow &window)
